+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/key/DefaultKeyboardShortcuts.java
@@ -15,13 +15,17 @@ import static java.util.stream.Collectors.toMap;
 
 final class DefaultKeyboardShortcuts<T extends Enum<T>> implements KeyboardShortcuts<T> {
 
+  private final Class<T> shortcutsClass;
   private final Map<T, Value<KeyStroke>> keyStrokes;
 
   DefaultKeyboardShortcuts(Class<T> shortcutsClass, Function<T, KeyStroke> defaultKeystrokes) {
-    requireNonNull(shortcutsClass);
-    requireNonNull(defaultKeystrokes);
-    keyStrokes = Stream.of(shortcutsClass.getEnumConstants())
-            .collect(toMap(Function.identity(), shortcutKey -> keyStrokeValue(defaultKeystrokes, shortcutKey)));
+    this(requireNonNull(shortcutsClass), Stream.of(shortcutsClass.getEnumConstants())
+            .collect(toMap(Function.identity(), shortcutKey -> keyStrokeValue(defaultKeystrokes, shortcutKey))));
+  }
+
+  private DefaultKeyboardShortcuts(Class<T> shortcutsClass, Map<T, Value<KeyStroke>> keyStrokes) {
+    this.shortcutsClass = shortcutsClass;
+    this.keyStrokes = keyStrokes;
   }
 
   /**
@@ -32,8 +36,15 @@ final class DefaultKeyboardShortcuts<T extends Enum<T>> implements KeyboardShort
     return keyStrokes.get(requireNonNull(keyboardShortcut));
   }
 
-  private Value<KeyStroke> keyStrokeValue(Function<T, KeyStroke> defaultKeystrokes, T shortcutKey) {
-    KeyStroke keyStroke = defaultKeystrokes.apply(shortcutKey);
+  @Override
+  public KeyboardShortcuts<T> copy() {
+    return new DefaultKeyboardShortcuts<>(shortcutsClass, keyStrokes.entrySet().stream()
+            .collect(toMap(Map.Entry::getKey, entry ->
+                    Value.value(entry.getValue().get(), entry.getValue().get()))));
+  }
+
+  private static <T extends Enum<T>> Value<KeyStroke> keyStrokeValue(Function<T, KeyStroke> defaultKeystrokes, T shortcutKey) {
+    KeyStroke keyStroke = requireNonNull(defaultKeystrokes).apply(shortcutKey);
     if (keyStroke == null) {
       throw new IllegalStateException("No default keystroke provided for shortcut key: " + shortcutKey);
     }