+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/key/KeyboardShortcutsTest.java
@@ -3,14 +3,17 @@
  */
 package is.codion.swing.common.ui.key;
 
+import is.codion.common.value.Value;
+
 import org.junit.jupiter.api.Test;
 
+import javax.swing.KeyStroke;
 import java.awt.event.KeyEvent;
+import java.util.stream.Stream;
 
 import static is.codion.swing.common.ui.key.KeyboardShortcuts.keyStroke;
 import static is.codion.swing.common.ui.key.KeyboardShortcuts.keyboardShortcuts;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.*;
 
 public final class KeyboardShortcutsTest {
 
@@ -41,5 +44,13 @@ public final class KeyboardShortcutsTest {
 
     assertEquals(KeyEvent.VK_1, shortcuts.keyStroke(Shortcut.ONE).get().getKeyCode());
     assertEquals(KeyEvent.VK_2, shortcuts.keyStroke(Shortcut.TWO).get().getKeyCode());
+
+    KeyboardShortcuts<Shortcut> copy = shortcuts.copy();
+    Stream.of(Shortcut.values()).forEach(shortcut -> {
+      Value<KeyStroke> keyStrokeValue = shortcuts.keyStroke(shortcut);
+      Value<KeyStroke> keyStrokeValueCopy = copy.keyStroke(shortcut);
+      assertNotSame(keyStrokeValue, keyStrokeValueCopy);
+      assertTrue(keyStrokeValue.equalTo(keyStrokeValueCopy.get()));
+    });
   }
 }